Many improvements in grbl message parser++
Some improvements needs my grblHAL port to function correctly as my
forked vanilla grbl lacks good support for MPG. Among lacking features
are message output synced to gcode execution and the fact that execution
continues after an error (until the sender stops transmitting and the
input buffer is empty - depends on streaming mode).  The latest grbHAL
version (to be updated soon) will return the last error until a reset, a
stop, a system command or a buffer flush is issued by the sender.